# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jewelry Store REST API",
    "description": "REST API theo database.sql 25 bảng. Đặt hàng COD; phân trang bắt đầu từ 0. KHACH_HANG, NHAN_VIEN, QUAN_LY. Xem README/API.md để biết phân quyền và luồng nghiệp vụ.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/inventory/import": {
      "post": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "add_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
    "/api/inventory/export": {
      "post": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "subtract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
    "/api/inventory/adjust": {
      "post": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "adjust",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
    "/api/gold-prices": {
      "get": {
        "tags": [
          "gold-price-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "history",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseGoldPriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      },
      "post": {
        "tags": [
          "gold-price-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoldPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGoldPriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
...
    "/api/inventory/{variantId}/threshold": {
      "patch": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "threshold",
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockThresholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
...
    "/api/inventory": {
      "get": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "list_10",
        "parameters": [
          {
            "name": "lowStock",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
    "/api/inventory/{variantId}": {
      "get": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "get_5",
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseInventoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
    "/api/inventory/{variantId}/history": {
      "get": {
        "tags": [
          "inventory-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: NHAN_VIEN, QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "history_1",
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseStockHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "NHAN_VIEN",
          "QUAN_LY"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoldPriceRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "buyPrice": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "sellPrice": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "unit": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          }
        },
        "required": [
          "buyPrice",
          "date",
          "sellPrice",
          "type",
          "unit"
        ]
      },
      "ApiResponseGoldPriceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/GoldPriceResponse"
          }
        }
      },
      "GoldPriceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "buyPrice": {
            "type": "number"
          },
          "sellPrice": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "unit": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InventoryRequest": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "maximum": 1000000,
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "reason",
          "variantId"
        ]
      },
      "ApiResponseInventoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/InventoryResponse"
          }
        }
      },
      "InventoryResponse": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "reserved": {
            "type": "integer",
            "format": "int32"
          },
          "available": {
            "type": "integer",
            "format": "int32"
          },
          "lowStock": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lowStockThreshold": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "StockThresholdRequest": {
        "type": "object",
        "properties": {
          "threshold": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": [
          "threshold"
        ]
      },
...
      "ApiResponsePageResponseInventoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseInventoryResponse"
          }
        }
      },
      "PageResponseInventoryResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponsePageResponseStockHistoryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseStockHistoryResponse"
          }
        }
      },
      "PageResponseStockHistoryResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockHistoryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "StockHistoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "variantId": {
            "type": "integer",
            "format": "int64"
          },
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "action": {
            "type": "string",
            "enum": [
              "DIEU_CHINH",
              "GIAI_PHONG",
              "GIAO_HANG",
              "GIU_HANG",
              "NHAP",
              "XUAT"
            ]
          },
          "quantityChange": {
            "type": "integer",
            "format": "int32"
          },
          "reservedChange": {
            "type": "integer",
            "format": "int32"
          },
          "quantityAfter": {
            "type": "integer",
            "format": "int32"
          },
          "reservedAfter": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponsePageResponseGoldPriceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseGoldPriceResponse"
          }
        }
      },
      "PageResponseGoldPriceResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoldPriceResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}